backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
motor
httpx
pydantic
python-jose[cryptography]
bcrypt
python-multipart
